fixing package dependencies
