+++ ellar/core/conf/app_settings_models.py
@@ -5,11 +5,11 @@ from pydantic.json import ENCODERS_BY_TYPE as encoders_by_type
 from starlette.exceptions import HTTPException as StarletteHTTPException
 from starlette.websockets import WebSocketClose
 
+from ellar.common_types import ASGIApp, TReceive, TScope, TSend
 from ellar.constants import DEFAULT_LOGGING as default_logging, LOG_LEVELS as log_levels
 from ellar.core.response import JSONResponse, PlainTextResponse
 from ellar.core.versioning import DefaultAPIVersioning
 from ellar.serializer import Serializer, SerializerFilter
-from ellar.types import ASGIApp, TReceive, TScope, TSend
 
 from .mixins import ConfigDefaultTypesMixin, TExceptionHandler, TMiddleware, TVersioning
