+++ ellar/core/routing/router/module.py
@@ -4,6 +4,7 @@ import uuid
 from starlette.routing import BaseRoute, Match, Mount as StarletteMount, Route, Router
 from starlette.types import ASGIApp
 
+from ellar.common_types import TReceive, TScope, TSend
 from ellar.compatible import AttributeDict
 from ellar.constants import (
     CONTROLLER_CLASS_KEY,
@@ -16,16 +17,15 @@ from ellar.constants import (
 )
 from ellar.core.controller import ControllerBase
 from ellar.core.routing.route import RouteOperation
-from ellar.core.schema import RouteParameters, WsRouteParameters
 from ellar.helper import get_unique_control_type
 from ellar.reflect import reflect
-from ellar.types import TReceive, TScope, TSend
 
 from ..operation_definitions import (
     OperationDefinitions,
     TOperation,
     TWebsocketOperation,
 )
+from ..schema import RouteParameters, WsRouteParameters
 from .route_collections import RouteCollection
 
 if t.TYPE_CHECKING:  # pragma: no cover
