+++ ellar/testing/module.py
@@ -5,12 +5,12 @@ from uuid import uuid4
 from starlette.testclient import TestClient as TestClient
 
 from ellar.common import Module
+from ellar.common_types import T
 from ellar.core import ModuleBase
 from ellar.core.factory import AppFactory
 from ellar.core.main import App
 from ellar.core.routing import ModuleRouter
 from ellar.di import ProviderConfig
-from ellar.types import T
 
 if t.TYPE_CHECKING:  # pragma: no cover
     from ellar.core import GuardCanActivate
