+++ mypy.ini
@@ -27,9 +27,6 @@ ignore_errors = True
 [mypy-ellar.cli.*]
 ignore_errors = True
 
-[mypy-ellar.core.schema.*]
-ignore_errors = True
-
 [mypy-ellar.__main__.*]
 ignore_errors = True
