+++ tests/test_reflector.py
@@ -1,5 +1,5 @@
+from ellar.core.services import Reflector
 from ellar.reflect import reflect
-from ellar.services.reflector import Reflector
 
 reflector = Reflector()
