+++ tests/test_routing/sample.py
@@ -4,20 +4,20 @@ from typing import List, Optional
 
 from pydantic import Field
 
-from ellar.core.schema import Schema
+from ellar.serializer import Serializer
 
 
-class Product(Schema):
+class Product(Serializer):
     name: str
     description: str = None  # type: ignore
     price: float
 
 
-class Item(Schema):
+class Item(Serializer):
     name: Optional[str] = None
 
 
-class OtherItem(Schema):
+class OtherItem(Serializer):
     price: int
 
 
@@ -27,24 +27,24 @@ class Range(IntEnum):
     TWO_HUNDRED = 200
 
 
-class Filter(Schema):
+class Filter(Serializer):
     to_datetime: datetime = Field(alias="to")
     from_datetime: datetime = Field(alias="from")
     range: Range = Range.TWENTY
 
 
-class Data(Schema):
+class Data(Serializer):
     an_int: int = Field(alias="int", default=0)
     a_float: float = Field(alias="float", default=1.5)
 
 
-class NonPrimitiveSchema(Schema):
+class NonPrimitiveSchema(Serializer):
     # The schema can only work for Body
     # And will fail for Path, Cookie, Query, Header, Form
     filter: Filter
 
 
-class ListOfPrimitiveSchema(Schema):
+class ListOfPrimitiveSchema(Serializer):
     # The schema can only work for Body, Query, Header and Form
     # And will fail for Path, Cookie
     an_int: List[int] = Field(alias="int")
